+++ pkg/controller/nodes/dynamic/handler_test.go
@@ -5,8 +5,6 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/lyft/flyteidl/gen/pb-go/flyteidl/admin"
-
 	"github.com/lyft/flytestdlib/contextutils"
 	"github.com/lyft/flytestdlib/promutils/labeled"
 
@@ -608,230 +606,6 @@ func createDynamicJobSpecWithLaunchPlans() *core.DynamicJobSpec {
 			},
 		},
 	}
-}
-
-func Test_dynamicNodeHandler_buildContextualDynamicWorkflow_withLaunchPlans(t *testing.T) {
-	createNodeContext := func(ttype string, finalOutput storage.DataReference) *nodeMocks.NodeExecutionContext {
-		ctx := context.TODO()
-
-		wfExecID := &core.WorkflowExecutionIdentifier{
-			Project: "project",
-			Domain:  "domain",
-			Name:    "name",
-		}
-
-		nodeID := "n1"
-		nm := &nodeMocks.NodeExecutionMetadata{}
-		nm.OnGetAnnotations().Return(map[string]string{})
-		nm.OnGetNodeExecutionID().Return(&core.NodeExecutionIdentifier{ExecutionId: wfExecID, NodeId: nodeID})
-		nm.OnGetK8sServiceAccount().Return("service-account")
-		nm.OnGetLabels().Return(map[string]string{})
-		nm.OnGetNamespace().Return("namespace")
-		nm.OnGetOwnerID().Return(types.NamespacedName{Namespace: "namespace", Name: "name"})
-		nm.OnGetOwnerReference().Return(v1.OwnerReference{
-			Kind: "sample",
-			Name: "name",
-		})
-
-		taskID := &core.Identifier{}
-		tk := &core.TaskTemplate{
-			Id:   taskID,
-			Type: "test",
-			Metadata: &core.TaskMetadata{
-				Discoverable: true,
-			},
-			Interface: &core.TypedInterface{
-				Outputs: &core.VariableMap{
-					Variables: map[string]*core.Variable{
-						"x": {
-							Type: &core.LiteralType{
-								Type: &core.LiteralType_Simple{
-									Simple: core.SimpleType_INTEGER,
-								},
-							},
-						},
-					},
-				},
-			},
-		}
-		tr := &nodeMocks.TaskReader{}
-		tr.OnGetTaskID().Return(taskID)
-		tr.OnGetTaskType().Return(ttype)
-		tr.OnReadMatch(mock.Anything).Return(tk, nil)
-
-		n := &flyteMocks.ExecutableNode{}
-		tID := "dyn-task-1"
-		n.OnGetTaskID().Return(&tID)
-
-		dataStore, err := storage.NewDataStore(&storage.Config{Type: storage.TypeMemory}, promutils.NewTestScope())
-		assert.NoError(t, err)
-
-		ir := &ioMocks.InputReader{}
-		nCtx := &nodeMocks.NodeExecutionContext{}
-		nCtx.OnNodeExecutionMetadata().Return(nm)
-		nCtx.OnNode().Return(n)
-		nCtx.OnInputReader().Return(ir)
-		nCtx.OnCurrentAttempt().Return(uint32(1))
-		nCtx.OnTaskReader().Return(tr)
-		nCtx.OnMaxDatasetSizeBytes().Return(int64(1))
-		nCtx.OnNodeID().Return("n1")
-		nCtx.OnEnqueueOwnerFunc().Return(func() error { return nil })
-		nCtx.OnDataStore().Return(dataStore)
-
-		endNodeStatus := &flyteMocks.ExecutableNodeStatus{}
-		endNodeStatus.OnGetDataDir().Return(storage.DataReference("end-node"))
-		endNodeStatus.OnGetOutputDir().Return(storage.DataReference("end-node"))
-
-		subNs := &flyteMocks.ExecutableNodeStatus{}
-		subNs.On("SetDataDir", mock.Anything).Return()
-		subNs.On("SetOutputDir", mock.Anything).Return()
-		subNs.On("ResetDirty").Return()
-		subNs.OnGetOutputDir().Return(finalOutput)
-		subNs.On("SetParentTaskID", mock.Anything).Return()
-		subNs.OnGetAttempts().Return(0)
-
-		dynamicNS := &flyteMocks.ExecutableNodeStatus{}
-		dynamicNS.On("SetDataDir", mock.Anything).Return()
-		dynamicNS.On("SetOutputDir", mock.Anything).Return()
-		dynamicNS.On("SetParentTaskID", mock.Anything).Return()
-		dynamicNS.OnGetNodeExecutionStatus(ctx, "n1-1-Node_1").Return(subNs)
-		dynamicNS.OnGetNodeExecutionStatus(ctx, v1alpha1.EndNodeID).Return(endNodeStatus)
-
-		ns := &flyteMocks.ExecutableNodeStatus{}
-		ns.OnGetDataDir().Return(storage.DataReference("data-dir"))
-		ns.OnGetOutputDir().Return(storage.DataReference("output-dir"))
-		ns.OnGetNodeExecutionStatus(ctx, dynamicNodeID).Return(dynamicNS)
-		ns.OnGetNodeExecutionStatus(ctx, dynamicNodeID).Return(dynamicNS)
-		nCtx.OnNodeStatus().Return(ns)
-
-		w := &flyteMocks.ExecutableWorkflow{}
-		ws := &flyteMocks.ExecutableWorkflowStatus{}
-		ws.OnGetNodeExecutionStatus(ctx, "n1").Return(ns)
-		w.OnGetExecutionStatus().Return(ws)
-
-		r := &nodeMocks.NodeStateReader{}
-		r.OnGetDynamicNodeState().Return(handler.DynamicNodeState{
-			Phase: v1alpha1.DynamicNodePhaseExecuting,
-		})
-		nCtx.OnNodeStateReader().Return(r)
-		execContext := &executorMocks.ExecutionContext{}
-		nCtx.OnExecutionContext().Return(execContext)
-		return nCtx
-	}
-
-	t.Run("launch plan interfaces match parent task interface", func(t *testing.T) {
-		ctx := context.Background()
-		lpID := &core.Identifier{
-			ResourceType: core.ResourceType_LAUNCH_PLAN,
-			Name:         "my_plan",
-			Project:      "p",
-			Domain:       "d",
-		}
-		djSpec := createDynamicJobSpecWithLaunchPlans()
-		finalOutput := storage.DataReference("/subnode")
-		nCtx := createNodeContext("test", finalOutput)
-		s := &dynamicNodeStateHolder{}
-		nCtx.On("NodeStateWriter").Return(s)
-		f, err := nCtx.DataStore().ConstructReference(ctx, nCtx.NodeStatus().GetOutputDir(), "futures.pb")
-		assert.NoError(t, err)
-		assert.NoError(t, nCtx.DataStore().WriteProtobuf(context.TODO(), f, storage.Options{}, djSpec))
-
-		mockLPLauncher := &lpMocks.Reader{}
-		var callsAdmin = false
-		mockLPLauncher.OnGetLaunchPlanMatch(ctx, lpID).Run(func(args mock.Arguments) {
-			// When a launch plan node is detected, a call should be made to Admin to fetch the interface for the LP
-			callsAdmin = true
-		}).Return(&admin.LaunchPlan{
-			Id: lpID,
-			Closure: &admin.LaunchPlanClosure{
-				ExpectedInputs: &core.ParameterMap{},
-				ExpectedOutputs: &core.VariableMap{
-					Variables: map[string]*core.Variable{
-						"x": {
-							Type: &core.LiteralType{
-								Type: &core.LiteralType_Simple{
-									Simple: core.SimpleType_INTEGER,
-								},
-							},
-							Description: "output of the launch plan",
-						},
-					},
-				},
-			},
-		}, nil)
-		h := &mocks.TaskNodeHandler{}
-		n := &executorMocks.Node{}
-		d := dynamicNodeTaskNodeHandler{
-			TaskNodeHandler: h,
-			nodeExecutor:    n,
-			lpReader:        mockLPLauncher,
-			metrics:         newMetrics(promutils.NewTestScope()),
-		}
-		execCtx, executableWorkflow, nl, isDynamic, err := d.buildContextualDynamicWorkflow(ctx, nCtx)
-		assert.True(t, callsAdmin)
-		assert.True(t, isDynamic)
-		assert.NoError(t, err)
-		assert.NotNil(t, executableWorkflow)
-		assert.NotNil(t, execCtx)
-		assert.NotNil(t, nl)
-	})
-
-	t.Run("launch plan interfaces do not parent task interface", func(t *testing.T) {
-		ctx := context.Background()
-		lpID := &core.Identifier{
-			ResourceType: core.ResourceType_LAUNCH_PLAN,
-			Name:         "my_plan",
-			Project:      "p",
-			Domain:       "d",
-		}
-		djSpec := createDynamicJobSpecWithLaunchPlans()
-		finalOutput := storage.DataReference("/subnode")
-		nCtx := createNodeContext("test", finalOutput)
-		s := &dynamicNodeStateHolder{}
-		nCtx.OnNodeStateWriter().Return(s)
-		f, err := nCtx.DataStore().ConstructReference(ctx, nCtx.NodeStatus().GetOutputDir(), "futures.pb")
-		assert.NoError(t, err)
-		assert.NoError(t, nCtx.DataStore().WriteProtobuf(context.TODO(), f, storage.Options{}, djSpec))
-
-		mockLPLauncher := &lpMocks.Reader{}
-		var callsAdmin = false
-		mockLPLauncher.OnGetLaunchPlanMatch(ctx, lpID).Run(func(args mock.Arguments) {
-			// When a launch plan node is detected, a call should be made to Admin to fetch the interface for the LP
-			callsAdmin = true
-		}).Return(&admin.LaunchPlan{
-			Id: lpID,
-			Closure: &admin.LaunchPlanClosure{
-				ExpectedInputs: &core.ParameterMap{},
-				ExpectedOutputs: &core.VariableMap{
-					Variables: map[string]*core.Variable{
-						"d": {
-							Type: &core.LiteralType{
-								Type: &core.LiteralType_Simple{
-									Simple: core.SimpleType_STRING,
-								},
-							},
-							Description: "output of the launch plan",
-						},
-					},
-				},
-			},
-		}, nil)
-		h := &mocks.TaskNodeHandler{}
-		n := &executorMocks.Node{}
-		d := dynamicNodeTaskNodeHandler{
-			TaskNodeHandler: h,
-			nodeExecutor:    n,
-			lpReader:        mockLPLauncher,
-			metrics:         newMetrics(promutils.NewTestScope()),
-		}
-		execCtx, executableWorkflow, nl, isDynamic, err := d.buildContextualDynamicWorkflow(ctx, nCtx)
-		assert.True(t, callsAdmin)
-		assert.True(t, isDynamic)
-		assert.Error(t, err)
-		assert.Nil(t, executableWorkflow)
-		assert.Nil(t, execCtx)
-		assert.Nil(t, nl)
-	})
 }
 
 func TestDynamicNodeTaskNodeHandler_Finalize(t *testing.T) {
