+++ pkg/controller/nodes/transformers.go
@@ -144,6 +144,7 @@ func UpdateNodeStatus(np v1alpha1.NodePhase, p handler.PhaseInfo, n *nodeStateMa
 		t := s.GetOrCreateDynamicNodeStatus()
 		t.SetDynamicNodePhase(n.d.Phase)
 		t.SetDynamicNodeReason(n.d.Reason)
+		t.SetExecutionError(n.d.Error)
 	}
 
 	// Update branch node status
